get, add category. cal avg rate

+++ migrations/20230205053935-create-product.js
@@ -18,7 +18,7 @@ module.exports = {
         allowNull: false,
       },
       rate: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.FLOAT,
       },
       price: {
         type: Sequelize.DECIMAL,
